Split Query::ClassMethods in two and add comments, rename `query` `scope`
