+++ app/classes/query/modules/subqueries.rb
@@ -2,67 +2,19 @@
 
 # Methods that are available to instances as class methods, and to ::Query.
 # ::Query is a convenience delegator class so callers can access these methods.
-module Query::Modules::ClassMethods
+#
+#  Subqueries:
+#  current_or_related_query:: Convert queries from one model to another; can be
+#                             called recursively. To avoid repetitive recursion,
+#                             it checks for a nested query that may be for the
+#                             intended target model.
+#
+module Query::Modules::Subqueries
   def self.included(base)
     base.extend(ClassMethods)
   end
 
   module ClassMethods
-    # Query::Modules::ActiveRecord
-    def safe_find(id)
-      find(id)
-    rescue ::ActiveRecord::RecordNotFound
-      nil
-    end
-
-    def find(id)
-      record = QueryRecord.find(id)
-      query = Query.rebuild_from_description(record.description)
-      record.query = query
-      query.record = record
-      QueryRecord.cleanup
-      query
-    end
-
-    def lookup_and_save(*)
-      query = lookup(*)
-      query.record.save!
-      query
-    end
-
-    def lookup(*)
-      query = Query.new(*)
-      record = get_record(query)
-      record.query = query
-      query.record = record
-      QueryRecord.cleanup
-      query
-    end
-
-    def get_record(query)
-      desc = query.serialize
-      QueryRecord.find_by(description: desc) ||
-        QueryRecord.new(
-          description: desc,
-          updated_at: Time.zone.now,
-          access_count: 0
-        )
-    end
-
-    # Query::Modules::Serialization
-    #
-    # Get the model from the serialized params and instantiate new Query.
-    def rebuild_from_description(description)
-      model, params = deserialize(description)
-      ::Query.new(model, params)
-    end
-
-    def deserialize(description)
-      params = JSON.parse(description).deep_symbolize_keys
-      model = params.delete(:model)
-      [model, params]
-    end
-
     # Query::Modules::RelatedQueries
     #
     # Query needs to know which joins are necessary to make these conversions
