+++ app/controllers/names/maps_controller.rb
@@ -13,7 +13,7 @@ module Names
 
       @query = find_or_create_query(:Observation, names: { lookup: @name.id })
       @any_content_filters_applied = check_if_preference_filters_applied
-      @observations = @query.query.limit(MO.query_max_array).
+      @observations = @query.scope.limit(MO.query_max_array).
                       includes(:location).
                       select { |o| o.lat || o.location }
     end
